# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: Nabiha Financial Intelligence Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /expenses:
    get:
      operationId: listExpenses
      tags: [expenses]
      summary: List all expenses
      parameters:
        - name: categoryId
          in: query
          required: false
          schema:
            type: integer
        - name: priority
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/Priority"
        - name: month
          in: query
          required: false
          schema:
            type: string
            description: "YYYY-MM format"
      responses:
        "200":
          description: List of expenses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Expense"
    post:
      operationId: createExpense
      tags: [expenses]
      summary: Create a new expense
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateExpenseBody"
      responses:
        "201":
          description: Created expense
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
  /expenses/{id}:
    get:
      operationId: getExpense
      tags: [expenses]
      summary: Get expense by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Expense
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
        "404":
          description: Not found
    put:
      operationId: updateExpense
      tags: [expenses]
      summary: Update an expense
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateExpenseBody"
      responses:
        "200":
          description: Updated expense
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
    delete:
      operationId: deleteExpense
      tags: [expenses]
      summary: Delete an expense
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all categories
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create a new category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryBody"
      responses:
        "201":
          description: Created category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /categories/{id}:
    delete:
      operationId: deleteCategory
      tags: [categories]
      summary: Delete a category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /commitments:
    get:
      operationId: listCommitments
      tags: [commitments]
      summary: List all financial commitments
      responses:
        "200":
          description: List of commitments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Commitment"
    post:
      operationId: createCommitment
      tags: [commitments]
      summary: Create a financial commitment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCommitmentBody"
      responses:
        "201":
          description: Created commitment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Commitment"
  /commitments/{id}:
    put:
      operationId: updateCommitment
      tags: [commitments]
      summary: Update a commitment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCommitmentBody"
      responses:
        "200":
          description: Updated commitment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Commitment"
    delete:
      operationId: deleteCommitment
      tags: [commitments]
      summary: Delete a commitment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /events:
    get:
      operationId: listEvents
      tags: [events]
      summary: List financial calendar events
      parameters:
        - name: year
          in: query
          required: false
          schema:
            type: integer
        - name: month
          in: query
          required: false
          schema:
            type: string
            description: "YYYY-MM format"
      responses:
        "200":
          description: Events list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Event"
    post:
      operationId: createEvent
      tags: [events]
      summary: Create an event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEventBody"
      responses:
        "201":
          description: Created event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
  /events/{id}:
    delete:
      operationId: deleteEvent
      tags: [events]
      summary: Delete an event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /notifications/{id}/read:
    post:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Marked as read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
...
components:
  schemas:
...
    Priority:
      type: string
      enum: [essential, important, luxury]
...
    Expense:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        amount:
          type: number
        priority:
          $ref: "#/components/schemas/Priority"
        categoryId:
          type: integer
          nullable: true
        categoryName:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        date:
          type: string
          format: date
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - amount
        - priority
        - date
        - createdAt
    CreateExpenseBody:
      type: object
      properties:
        title:
          type: string
        amount:
          type: number
        priority:
          $ref: "#/components/schemas/Priority"
        categoryId:
          type: integer
          nullable: true
        notes:
          type: string
          nullable: true
        date:
          type: string
          format: date
      required:
        - title
        - amount
        - priority
        - date
    UpdateExpenseBody:
      type: object
      properties:
        title:
          type: string
        amount:
          type: number
        priority:
          $ref: "#/components/schemas/Priority"
        categoryId:
          type: integer
          nullable: true
        notes:
          type: string
          nullable: true
        date:
          type: string
          format: date
    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        icon:
          type: string
          nullable: true
        color:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - createdAt
    CreateCategoryBody:
      type: object
      properties:
        name:
          type: string
        icon:
          type: string
          nullable: true
        color:
          type: string
          nullable: true
      required:
        - name
    Commitment:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        amount:
          type: number
        dueDay:
          type: integer
        isPaid:
          type: boolean
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - amount
        - dueDay
        - isPaid
        - createdAt
    CreateCommitmentBody:
      type: object
      properties:
        title:
          type: string
        amount:
          type: number
        dueDay:
          type: integer
        notes:
          type: string
          nullable: true
      required:
        - title
        - amount
        - dueDay
    UpdateCommitmentBody:
      type: object
      properties:
        title:
          type: string
        amount:
          type: number
        dueDay:
          type: integer
        isPaid:
          type: boolean
        notes:
          type: string
          nullable: true
...
    EventType:
      type: string
      enum: [salary, bill, subscription, goal, reminder, expense]
    Event:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        amount:
          type: number
          nullable: true
        date:
          type: string
          format: date
        type:
          $ref: "#/components/schemas/EventType"
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - date
        - type
        - createdAt
    CreateEventBody:
      type: object
      properties:
        title:
          type: string
        amount:
          type: number
          nullable: true
        date:
          type: string
          format: date
        type:
          $ref: "#/components/schemas/EventType"
        notes:
          type: string
          nullable: true
      required:
        - title
        - date
        - type
    NotificationType:
      type: string
      enum: [info, warning, success, danger]
    Notification:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        message:
          type: string
        type:
          $ref: "#/components/schemas/NotificationType"
        isRead:
          type: boolean
        link:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - message
        - type
        - isRead
        - createdAt
...